Use static gas price adapter in test harness
