+++ crates/fuel-core/src/service/adapters/producer.rs
@@ -12,13 +12,7 @@ use crate::{
     },
 };
 use fuel_core_executor::executor::OnceTransactionsSource;
-use fuel_core_producer::{
-    block_producer::gas_price::{
-        GasPriceParams,
-        ProducerGasPrice,
-    },
-    ports::TxPool,
-};
+use fuel_core_producer::ports::TxPool;
 use fuel_core_storage::{
     not_found,
     tables::FuelBlocks,
@@ -144,21 +138,5 @@ impl fuel_core_producer::ports::BlockProducerDatabase for Database {
 
     fn block_header_merkle_root(&self, height: &BlockHeight) -> StorageResult<Bytes32> {
         self.storage::<FuelBlocks>().root(height).map(Into::into)
-    }
-}
-
-pub struct StaticGasPrice {
-    pub gas_price: u64,
-}
-
-impl StaticGasPrice {
-    pub fn new(gas_price: u64) -> Self {
-        Self { gas_price }
-    }
-}
-
-impl ProducerGasPrice for StaticGasPrice {
-    fn gas_price(&self, _params: GasPriceParams) -> u64 {
-        self.gas_price
     }
 }
