+++ crates/services/producer/src/block_producer/tests.rs
@@ -1,5 +1,8 @@
 use crate::{
-    block_producer::Error,
+    block_producer::{
+        gas_price::StaticGasPrice,
+        Error,
+    },
     mocks::{
         FailingMockExecutor,
         MockDb,
@@ -251,7 +254,8 @@ impl<Executor> TestContext<Executor> {
         }
     }
 
-    pub fn producer(self) -> Producer<MockDb, MockTxPool, Executor> {
+    pub fn producer(self) -> Producer<MockDb, MockTxPool, Executor, StaticGasPrice> {
+        let gas_price = self.config.gas_price;
         Producer {
             config: self.config,
             view_provider: self.db,
@@ -259,6 +263,7 @@ impl<Executor> TestContext<Executor> {
             executor: self.executor,
             relayer: Box::new(self.relayer),
             lock: Default::default(),
+            gas_price_provider: StaticGasPrice::new(gas_price),
         }
     }
 }
